% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cs.R
\name{plot_cs}
\alias{plot_cs}
\title{Plot and save a compound simetry matrix with dimension equals to n}
\usage{
plot_cs(n = 5, rho = 0.5, save_path = NULL, width = 5, height = 5, dpi = 300)
}
\arguments{
\item{n}{size of the matrix}

\item{rho}{correlation}

\item{save_path}{Path to save the chart (.png). If NULL, do not save.}

\item{width}{Width of the saved chart, in inches.}

\item{height}{Height of the saved chart, in inches.}

\item{dpi}{Resolution of the saved chart (dots per inch)}
}
\value{
a figure
}
\description{
Plot and save a compound simetry matrix with dimension equals to n
}
\examples{
plot_cs(n=5, rho=0.5, save_path = "plot.png", width = 8, height = 8, dpi = 300)

}
