% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fa.R
\name{plot_fa}
\alias{plot_fa}
\title{Plot and save a kth order factor analityc matrix with dimension equals to n}
\usage{
plot_fa(n = 5, k = 1, save_path = NULL, width = 5, height = 5, dpi = 300)
}
\arguments{
\item{n}{size of the matrix}

\item{k}{order of FA}

\item{save_path}{Caminho para salvar o gráfico (.png). Se NULL, não salva.}

\item{width}{Largura do gráfico salvo, em polegadas.}

\item{height}{Altura do gráfico salvo, em polegadas.}

\item{dpi}{Resolução do gráfico salvo (dots per inch)}
}
\value{
a figure
}
\description{
Plot and save a kth order factor analityc matrix with dimension equals to n
}
\examples{
plot_fa(n=5, k = 1, save_path = "plot.png", width = 8, height = 8, dpi = 300)

}
