% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_identity.R
\name{plot_identity}
\alias{plot_identity}
\title{Plot a identity matrix with dimension equals to n}
\usage{
plot_identity(n = 5, save_path = NULL, width = 5, height = 5, dpi = 300)
}
\arguments{
\item{n}{}
}
\description{
Plot a identity matrix with dimension equals to n
}
\examples{
plot_identity(n=5)

}
