% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_us.R
\name{plot_us}
\alias{plot_us}
\title{Plot and save a unstructured matrix with dimension equals to n}
\usage{
plot_us(n = 5, save_path = NULL, width = 5, height = 5, dpi = 300)
}
\arguments{
\item{n}{dimension of the matrix}

\item{save_path}{Path to save the chart (.png). If NULL, do not save.}

\item{width}{Width of the saved chart, in inches.}

\item{height}{Height of the saved chart, in inches.}

\item{dpi}{Resolution of the saved chart (dots per inch)}
}
\value{
a figure
}
\description{
Plot and save a unstructured matrix with dimension equals to n
}
\examples{
plot_us(n=5, k = 1, save_path = "plot.png", width = 8, height = 8, dpi = 300)

}
