% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fa.R
\name{plot_fa}
\alias{plot_fa}
\title{Plot and save a kth order factor analytic matrix scaled to max = 1}
\usage{
plot_fa(n = 5, k = 1, save_path = NULL, width = 5, height = 5, dpi = 300)
}
\arguments{
\item{n}{size of the matrix}

\item{k}{order of FA}

\item{save_path}{Path to save the chart (.png). If NULL, do not save.}

\item{width}{Width of the saved chart, in inches.}

\item{height}{Height of the saved chart, in inches.}

\item{dpi}{Resolution of the saved chart (dots per inch)}
}
\value{
a figure
}
\description{
Plot and save a kth order factor analytic matrix scaled to max = 1
}
\examples{
plot_fa(n=5, k = 1, save_path = "plot.png", width = 8, height = 8, dpi = 300)
}
