% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_power_het.R
\name{plot_power_het}
\alias{plot_power_het}
\title{Plot and save a heterogeneous power matrix with dimension equals to n}
\usage{
plot_power_het(
  n = 5,
  rho = 0.7,
  save_path = NULL,
  width = 5,
  height = 5,
  dpi = 300
)
}
\arguments{
\item{n}{size of the matrix}

\item{rho}{correlation}

\item{save_path}{Path to save the chart (.png). If NULL, do not save.}

\item{width}{Width of the saved chart, in inches.}

\item{height}{Height of the saved chart, in inches.}

\item{dpi}{Resolution of the saved chart (dots per inch)}
}
\value{
a figure
}
\description{
Plot and save a heterogeneous power matrix with dimension equals to n
}
\examples{
plot_power_het(n=5, rho = 0.7, save_path = "plot.png", width = 8, height = 8, dpi = 300)

}
